+++ decidim-meetings/spec/system/meeting_registrations_spec.rb
@@ -61,7 +61,7 @@ describe "Meeting registrations", type: :system do
 
         within ".card.extra" do
           expect(page).to have_css("button[disabled]", text: "NO SLOTS AVAILABLE")
-          expect(page).to have_text("No slots remaining")
+          expect(page).to have_text("0 slots remaining")
         end
       end
     end