+++ spec/i18n_spec.rb
@@ -10,7 +10,6 @@ describe "I18n sanity" do
   let(:i18n) { I18n::Tasks::BaseTask.new(locales: locales.split(",")) }
   let(:missing_keys) { i18n.missing_keys }
   let(:unused_keys) { i18n.unused_keys }
-  let(:non_normalized_paths) { i18n.non_normalized_paths }
 
   it "does not have missing keys" do
     expect(missing_keys).to be_empty, "#{missing_keys.inspect} are missing"
@@ -18,15 +17,5 @@ describe "I18n sanity" do
 
   it "does not have unused keys" do
     expect(unused_keys).to be_empty, "#{unused_keys.inspect} are unused"
-  end
-
-  unless ENV["SKIP_NORMALIZATION"]
-    it "is normalized" do
-      error_message = "The following files need to be normalized:\n" \
-                      "#{non_normalized_paths.map { |path| "  #{path}" }.join("\n")}\n" \
-                      "Please run `bundle exec i18n-tasks normalize --locales #{locales}` to fix them"
-
-      expect(non_normalized_paths).to be_empty, error_message
-    end
   end
 end